tests/wpt/web-platform-tests/content-security-policy/navigate-to/form-action/form-action-blocks-navigate-to-allows.html

<!DOCTYPE html>

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  var t = async_test("Test that form-action overrides navigate-to when present.");
  window.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.result, 'fail');
    assert_equals(e.data.violatedDirective, 'form-action');
  });
</script>
<iframe src="../support/form_action_navigation.sub.html?csp=navigate-to%20%27self%27%3B%20form-action%20%27none%27%3B&action=post_message_to_frame_owner.html&report_id=dummy">
</body>
